refactor(sync-v2): rename SyncAgent module from manager to agent
