+++ hathor/p2p/sync_v2/mempool.py
@@ -21,7 +21,7 @@ from twisted.internet.defer import Deferred, inlineCallbacks
 from hathor.transaction import BaseTransaction
 
 if TYPE_CHECKING:
-    from hathor.p2p.sync_v2.manager import NodeBlockSync
+    from hathor.p2p.sync_v2.agent import NodeBlockSync
 
 logger = get_logger()
